refactor initial folder and file structure
