--- pettopia-frontend/src/HomePage/HomePage.jsx
@@ -1,7 +1,0 @@
-export default function HomePage() {
-    return (
-        <>
-            <h3>Testing Main Page</h3>
-        </>
-    )
-}
