--- pettopia-frontend/src/HomePage/NavBar.jsx
@@ -1,9 +1,0 @@
-export default function NavBar() {
-    return (
-        <>
-            <nav>
-                <h1>Testing</h1>
-            </nav>
-        </>
-    )
-}
