--- pettopia-frontend/src/SearchPage/SearchPage.jsx
@@ -1,7 +1,0 @@
-export default function SearchPage() {
-    return (
-    <>
-        <h1>Testing Search Page</h1>
-    </>
-    )
-}
