+++ pettopia-frontend/src/pages/HomePage/HomePage.jsx
@@ -1,0 +1,10 @@
+export default function HomePage() {
+  return (
+    <>
+      <h3>Testing Main Page</h3>
+      <h1 className='text-3xl font-bold'>
+        Testing Tailwind CSS - Hello world!
+      </h1>
+    </>
+  );
+}
