+++ pettopia-frontend/src/pages/ProductPage/ProductDetail.jsx
@@ -1,0 +1,10 @@
+export default function ProductDetail() {
+  return (
+    <>
+      <h1>
+        Testing out Product Page but this will be very different or might not
+        exist, but just incase
+      </h1>
+    </>
+  );
+}
