+++ pettopia-frontend/src/pages/SearchPage/SearchPage.jsx
@@ -1,0 +1,7 @@
+export default function SearchPage() {
+  return (
+    <>
+      <h1>Testing Search Page</h1>
+    </>
+  );
+}
